refactor(cost-insight): share gcs storage client pool setup
