+++ cost-insight/src/cost_insight/common/gcs_cache_references.py
@@ -4,6 +4,8 @@ from collections.abc import Callable, Iterable
 from concurrent.futures import ThreadPoolExecutor
 from dataclasses import dataclass
 
+from cost_insight.common.gcs_client import create_storage_client
+
 
 @dataclass(frozen=True)
 class AcReferenceExtraction:
@@ -38,7 +40,7 @@ def extract_action_cache_references_batch(
     if not names:
         return ()
 
-    client = _create_storage_client(project_id=project_id, pool_maxsize=max_workers)
+    client = create_storage_client(project_id=project_id, pool_maxsize=max_workers)
     bucket = client.bucket(bucket_name)
 
     def extract_one(ac_object_name: str) -> AcReferenceExtraction:
@@ -85,42 +87,6 @@ def extract_action_cache_references_batch(
 
     with ThreadPoolExecutor(max_workers=max_workers) as executor:
         return tuple(executor.map(extract_one, names))
-
-
-def _create_storage_client(*, project_id: str, pool_maxsize: int):
-    from google.cloud import storage
-
-    client = storage.Client(project=project_id)
-    _configure_storage_client_http_pool(client=client, pool_maxsize=pool_maxsize)
-    return client
-
-
-def _configure_storage_client_http_pool(*, client, pool_maxsize: int) -> None:
-    if pool_maxsize <= 0:
-        return
-
-    session = getattr(client, "_http", None)
-    if session is None:
-        return
-
-    adapters = getattr(session, "adapters", None)
-    if not isinstance(adapters, dict):
-        return
-
-    from requests.adapters import HTTPAdapter
-
-    for prefix in ("https://", "http://"):
-        current_adapter = adapters.get(prefix)
-        pool_connections = getattr(current_adapter, "_pool_connections", pool_maxsize)
-        max_retries = getattr(current_adapter, "max_retries", 0)
-        session.mount(
-            prefix,
-            HTTPAdapter(
-                pool_connections=pool_connections,
-                pool_maxsize=pool_maxsize,
-                max_retries=max_retries,
-            ),
-        )
 
 
 def extract_cas_references_from_action_result_bytes(
